+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.105
+
+
+- Updated 304
+
 # 0.104
 
 - Refactored immobilization